Move Imaging to core
